reorganized the js-code in  index.mjs file into many small js files and folders
