+++ js/api/constants.js
@@ -1,0 +1,9 @@
+// All API keys. Export so the can be imported in other modules
+export const API_BASE = "https://v2.api.noroff.dev";
+export const API_AUTH = "/auth";
+export const API_REGISTER = "/register";
+export const API_LOGIN = "/login";
+export const API_KEY_URL = "/create-api-key";
+export const API_POSTS = "/social/posts";
+
+export const API_KEY = "703903c0-1408-41b7-8149-20c22ed7aaa2";
