+++ js/routes/home.js
@@ -1,0 +1,5 @@
+import { setAuthListener } from "../ui/listeners/auth.js";
+
+export async function homePage() {
+	setAuthListener()
+}
