+++ js/ui/listeners/auth.js
@@ -1,0 +1,7 @@
+import { onAuth } from "../events/onAuth.js";
+
+export function setAuthListener() {
+	document.forms.auth.addEventListener("submit", onAuth);
+}
+
+// setAuthListener();
